Renamed userprofile component to edituserprofile component since this page will only be accessible by the authorized user.
