+++ src/components/containers/index.js
@@ -5,7 +5,7 @@ export { default as AboutPage } from "./AboutPageContainer";
 export { default as JobsPage } from "./JobsPageContainer";
 export { default as LoginPage } from "./LoginContainer.js";
 export { default as SignupPage } from "./SignupContainer";
-export { default as UserProfilePage } from "./UserProfilePageContainer";
+export { default as EditUserProfilePage } from "./EditUserProfilePageContainer";
 export { default as NavBar } from "./NavBarContainer";
 export { default as LinkEditContainer } from "./LinkEditContainer";
 export { default as LinkContainer } from "./LinkContainer"
