+++ src/components/views/index.js
@@ -7,6 +7,6 @@ export { default as AboutPageView } from "./AboutPageView";
 export { default as JobsPageView } from "./JobsPageView";
 export { default as LoginPageView } from "./LoginView";
 export { default as SignupPageView } from "./SignupView";
-export { default as UserProfilePageView } from "./UserProfilePageView";
+export { default as EditUserProfilePageView } from "./EditUserProfilePageView";
 export { default as LinkEditView } from "./LinkEditView";
 export { default as LinkView } from "./LinkView";
